+++ client/components/Home/Homepage.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { Breadcrumb } from 'antd';
+
+export default function Homepage() {
+  return (
+    <>
+      <Breadcrumb style={{ margin: '16px 0' }}>
+        <Breadcrumb.Item>Home</Breadcrumb.Item>
+        <Breadcrumb.Item>Detail</Breadcrumb.Item>
+      </Breadcrumb>
+      <div style={{ background: '#fff', padding: 24, minHeight: 380 }}>Content</div>
+    </>
+  );
+}
