+++ client/less/common.less
@@ -7,13 +7,6 @@ html, body, #__next, .layout.ant-layout {
   min-height: 700px !important;
 }
 
-.lang {
-  border-bottom: none !important;
-  &:hover {
-    border-bottom: none !important;
-  }
-}
-
 .margin__auto {
   margin: 0 auto !important;
 }
