+++ client/less/footer.less
@@ -76,27 +76,3 @@ footer {
     }
   }
 }
-.home{
-  footer {
-    .footer-wrap{
-      width: 100%;
-      padding: 0;
-      border-bottom: 1px solid rgba(255, 255, 255, 0.25);
-      .ant-row{
-        width: 100%;
-        max-width: 1200px;
-        padding: 86px 24px 93px 24px;
-        margin: auto;
-        
-      }
-    }
-    .bottom-bar{
-      margin: auto;
-      max-width: 1200px;
-      padding: 16px 24px;
-    }
-    .bottom-bar{
-      border-top: none;
-    }
-  }
-}
