--- client/less/home.less
@@ -1,98 +1,0 @@
-@import './antd-default.less';
-
-.home-page-wrapper {
-  width: 100%;
-  padding: 0;
-  overflow: hidden;
-  position: relative; 
-  // will-change: transform; // firefox 下渲染占用太大。
-  color: @home-text-color;
-  font-family: 'SF UI Display', "Helvetica Neue For Number", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "PingFang SC", "Hiragino Sans GB", "Microsoft YaHei", "Helvetica Neue", Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  .page {
-    width: 100%;
-    max-width: 1200px;
-    margin: auto;
-    position: relative;
-    padding: 0 24px;
-    h2 {
-      font-size: 38px;
-      line-height: 46px;
-      color: @home-text-color;
-      text-align: center;
-      font-weight: 400;
-      margin: 140px auto 100px;
-    }
-    h1,
-    h2,
-    h3,
-    h4,
-    p {
-      will-change: transform;
-    }
-  }
-}
-
-.banner-bg {
-  &-wrapper {
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    top: 0;
-    left: 0;
-    overflow: hidden
-  }
-  width: 120%;
-  height: 200px;
-  background: @home-bg-color;
-  position: absolute;
-  bottom: -200px;
-  left: 0%;
-  transform-origin: 0;
-  transform: rotate(-4deg);
-}
-
-.banner {
-  display: flex;
-  align-items: center;
-  height: 576px;
-  &-wrapper {
-    overflow: initial;
-    position: relative;
-    z-index: 1;
-  }
-  .text-wrapper {
-    width: 54%;
-    max-width: 600px;
-    min-height: 320px;
-    color: #0D1A26;
-    h1 {
-      font-size: 68px;
-      font-weight: 600;
-      line-height: 76px;
-      margin: 8px 0 28px;
-      letter-spacing: 0;
-      font-family: "Avenir-Heavy", 'SF UI Display', "Helvetica Neue For Number", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "PingFang SC", "Hiragino Sans GB", "Microsoft YaHei", "Helvetica Neue", Helvetica, Arial, sans-serif;
-    }
-    p {
-      font-size: 20px;
-      line-height: 40px;
-      color: @home-text-color;
-    }
-  }
-  .img-wrapper {
-    width: 46%;
-    max-width: 482px;
-    position: absolute;
-    right: 0;
-    bottom: 26px;
-  }
-}
-
-svg {
-  display: block;
-  g {
-    transform-origin: 50%;
-    transform-box: fill-box;
-  }
-}
