+++ client/less/styles.less
@@ -1,5 +1,4 @@
 @import './common.less';
 @import './header.less';
-@import './home.less';
 @import './footer.less';
 @import './responsive.less';
