+++ client/next.config.js
@@ -1,5 +1,4 @@
 /* eslint-disable */
-require('dotenv').config()
 const withLess = require('@zeit/next-less')
 const lessToJS = require('less-vars-to-js')
 const fs = require('fs')
@@ -15,6 +14,12 @@ module.exports = withLess({
   lessLoaderOptions: {
     javascriptEnabled: true,
     modifyVars: themeVariables, // make your antd custom effective
+  },
+  onDemandEntries: {
+    // period (in ms) where the server will keep pages in the buffer
+    maxInactiveAge: 25 * 1000,
+    // number of pages that should be kept simultaneously without being disposed
+    pagesBufferLength: 2,
   },
   webpack: (config, { isServer }) => {
     config.plugins = config.plugins || []
