+++ client/package.json
@@ -4,36 +4,32 @@
   "description": "",
   "main": "server.js",
   "scripts": {
-    "test": "jest",
+    "test": "npm run eslint && npm run jest",
     "dev:server": "nodemon server.js",
     "build": "next build",
     "start": "NODE_ENV=production node server.js",
     "dev": "next -p 3001",
-    "eslint": "eslint ./ --fix"
+    "eslint": "eslint ./ --fix",
+    "jest": "jest"
   },
   "dependencies": {
     "@feathersjs/client": "^3.7.8",
     "@zeit/next-less": "^1.0.1",
     "antd": "^3.22.0",
     "babel-plugin-import": "^1.12.0",
-    "cookie": "^0.4.0",
-    "dotenv": "^8.0.0",
     "dotenv-webpack": "^1.7.0",
-    "enquire-js": "^0.2.1",
     "express": "^4.17.1",
     "js-cookie": "^2.2.1",
     "less": "^3.10.1",
     "less-vars-to-js": "^1.3.0",
+    "moment": "^2.24.0",
     "next": "^9.0.2",
-    "next-i18next": "^0.52.1",
+    "next-i18next": "^1.2.1",
     "next-redux-saga": "^4.0.2",
-    "next-redux-wrapper": "^3.0.0-alpha.3",
+    "next-redux-wrapper": "^3.0.0",
     "next-routes": "^1.4.2",
     "null-loader": "^3.0.0",
     "prop-types": "^15.7.2",
-    "rc-queue-anim": "^1.7.0",
-    "rc-scroll-anim": "^2.5.9",
-    "rc-tween-one": "^2.5.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-redux": "^7.1.0",
