+++ client/pages/auth/login.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Login from '../../components/auth/Login';
+import Login from '../../components/Auth/Login';
 
 import '../../less/auth.less';
