+++ client/server.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const next = require('next');
 /* eslint-disable-next-line no-unused-vars */
-const dotenv = require('dotenv').config();
+// const dotenv = require('dotenv').config();
 const routes = require('./routes/routes')
 
 const nextI18NextMiddleware = require('next-i18next/middleware').default;
@@ -10,10 +10,7 @@ const nextI18next = require('./i18n');
 const dev = process.env.NODE_ENV !== 'production';
 const port = parseInt(process.env.PORT, 10) || 3001;
 
-const app = next({
-  // dir: './src',
-  dev
-});
+const app = next({ dev });
 
 const handler = routes.getRequestHandler(app)
